requirements.txt
streamlit==1.28.1
plotly==5.17.0
pandas==2.0.3
numpy==1.24.3
networkx==3.1
pyyaml==6.0
python-dateutil==2.8.2
scipy==1.11.2